Complete models package split
Renamed most multi-word files with underscores
Modified tests, utils, and views to function with new class locations
