+++ DataRepo/compositeviews.py
@@ -5,18 +5,10 @@ from django.apps import apps
 from django.db.models import F, Model
 
 from .models.animal import Animal
-from .models.peakdata import PeakData
-from .models.peakgroup import PeakGroup
+from .models.peak_data import PeakData
+from .models.peak_group import PeakGroup
 from .models.tissue import Tissue
-from .models.tracerlabeledclass import TracerLabeledClass
-
-#from DataRepo.models import (
-#    Animal,
-#    PeakData,
-#    PeakGroup,
-#    Tissue,
-#    TracerLabeledClass,
-#)
+from .models.tracer_labeled_class import TracerLabeledClass
 
 
 class BaseSearchView:
