+++ DataRepo/models/tissue.py
@@ -1,5 +1,6 @@
 from django.db import models
 
+
 class Tissue(models.Model):
     # Instance / model fields
     id = models.AutoField(primary_key=True)
@@ -22,4 +23,3 @@ class Tissue(models.Model):
 
     def __str__(self):
         return str(self.name)
-
