+++ DataRepo/models/tracer_labeled_class.py
