+++ DataRepo/tests/test_loading.py
@@ -1,7 +1,7 @@
 from django.core.management import CommandError, call_command
 from django.test import tag
 
-from DataRepo.models import Compound
+from DataRepo.models.compound import Compound
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
